etl_docker/requirements.txt
#requirements.txt

requests
psycopg2-binary
python-dotenv
pandas
sqlalchemy

